v1 complete

+++ admin/delete-category.php
@@ -6,15 +6,21 @@ if(isset($_GET['id'])){
 //for later
     //update catergory_id posts that belong to this category to if of uncategarised categpry
 
+    $update_query ="UPDATE posts SET category_id=9 WHERE category_id=$id";
+    $update_result= mysqli_query($connection, $update_query);
+
+    if (!mysqli_errno($connection)) {
 
 
 
 
-    //delete category
-    $query="DELETE FROM categories WHERE id=$id LIMIT 1";
-    $result = mysqli_query($connection, $query);
-    $_SESSION['delete-category-success'] = "category deleted successfully; ";
-
+        //delete category
+        $query = "DELETE FROM categories WHERE id=$id LIMIT 1";
+        $result = mysqli_query($connection, $query);
+        $_SESSION['delete-category-success'] = "category deleted successfully; ";
+    }
 }
 header('location: ' . ROOT_URL . 'admin/manage-categories.php');
 die();
+
+
